Fix stdout/stderr processing in lldb scripts
Seems that asynchronos commands cause LLDB to mess up stdout/stderr processing
and causes some of the output to be lost. This patch creates a separate thread
to process stdout/stderr in blocking fashion and disables processing by builtin
LLDB mechanisms. There is still a posibility for some output to be lost if LLDB
closes connection too early, but this is only fixable in debugserver itself.